+++ src/extension.ts
@@ -13,7 +13,7 @@ import * as text from "./text";
 
 import * as console from "./console";
 
-const EXTENSION_VERSION = "3.7.3";
+const EXTENSION_VERSION = "3.8.0";
 
 let newFileWatcher: vscode.FileSystemWatcher | undefined;
 let resetEventFileWatcher: vscode.FileSystemWatcher | undefined;
@@ -207,11 +207,11 @@ function createWatchers(fixableProject: Fixable) {
 		const fileName = filePath.base;
 
 		console.log(`\nWARNING: Detected new file (${fileName}).`);
-		console.log("WARNING: Restart VSCode to fix new file Intellisense errors.\n");
+		console.log("WARNING: Soft Reset Unreal project to fix new file Intellisense errors.\n");
 
 		console.outputChannel?.show(true);
 		
-		await vscode.window.showInformationMessage("New Soure/Header files detected: Restart VSCode if you get Intellisense errors with these new files.  This extension will add them to the appropriate file on startup.", text.OK);
+		await vscode.window.showInformationMessage("New Soure/Header files detected: Soft Reset your Unreal project and restart VSCode", text.OK);
 		
 	});
 }