+++ src/text.ts
@@ -1,6 +1,7 @@
 
 export const MESSAGE_BUILD_SUCCESSFUL_PROJECT = " You must build your project successfully for a fix to work correctly.\n Restart VSCode after a successful build.";
 export const OK = "OK";
+export const DONT_WARN_AGAIN = "Don't Warn Again";
 export const INVALID_REGEX = "Invalid Regular Expression. Preventing config change by throwing exception.";
 
 export const PLEASE_WAIT_INTELLISENSE_FIXES = "Please wait for Intellisense Fixes to run...";