+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ### [TODO]
 
+# [3.8.0] 2023-11-4
+### Changed
+- Now says to user to soft reset project which is proper way to add new source files intellisense.
+- Now uses esbuild to bundles extension (smaller download size and faster start)
+
 # [3.7.3] 2023-9-6
 ### Fixed
 - Cpp standard fix now supports UE 5.3+ and c++20